adjusted structure for packaging
