+++ comlocal/ComLocAL/.gitignore
@@ -1,0 +1,2 @@
+*.pyc
+*.egg-info
