+++ comlocal/ComLocAL/util/CommandPacket.py
