+++ setup.py
@@ -1,0 +1,11 @@
+from setuptools import setup
+
+setup(name='comlocal',
+      version='0.1',
+      description='Communication and Localization Abstraction Layer',
+      url='https://github.com/jtsiva/ComLocAL',
+      author='Josh Siva',
+      author_email='jtsiva@gmail.com',
+      license='MIT',
+      packages=['comlocal'],
+      zip_safe=False)
